+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix call to undefined method pushProcessor on handler that does not implement ProcessableHandlerInterface
 * Use "use_locking" option with rotating file handler
 * Add ability to specify custom Sentry hub service
+* Register channels to `ResetLoggersWorkerSubscriber`
 
 ## 3.6.0 (2020-10-06)